% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{email}
\alias{email}
\alias{email_settings}
\alias{list_email}
\title{Email}
\usage{
email_settings(...)

list_email(filter, ...)
}
\arguments{
\item{...}{Named parameters passed on to \code{\link[crul]{HttpClient}}}

\item{filter}{An email filter, one of received, sent, skipped, rejected}
}
\description{
Email
}
\examples{
\dontrun{
# email settings
email_settings()

# list settings for an email address
list_email("received")
}
}
