% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups}
\alias{groups}
\alias{group_members}
\alias{group_create}
\alias{group_delete}
\title{Work with groups}
\usage{
groups(...)

group_members(name, ...)

group_create(name, ...)

group_delete(id, ...)
}
\arguments{
\item{...}{Named parameters passed on to \code{\link[crul]{HttpClient}}}

\item{name}{(character) A group name. required}

\item{id}{(numeric) A group id. required}
}
\description{
Work with groups
}
\examples{
\dontrun{
# all groups
groups()

# specific group by name
group_members("admins")
group_members("moderators")
group_members("trust_level_3")

# create a group
(res <- group_create("group_testing2"))

# delete a group
group_delete(res$basic_group$id)
}
}
