% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_messages.R
\name{private_messages}
\alias{private_messages}
\title{Private messages}
\usage{
private_messages(username, url = NULL, key = NULL, user = NULL, ...)
}
\arguments{
\item{username}{A username. required.}

\item{url}{Base url for a Discourse installation. See \code{\link{disc_setup}}}

\item{key}{Your api key on the Discourse installation. See \code{\link{disc_setup}}}

\item{user}{Your user name on the Discourse installation. See \code{\link{disc_setup}}}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}},
\code{\link[httr]{POST}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}
}
\description{
Private messages
}
\examples{
\dontrun{
private_messages(username = "sckott")
}
}
