% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Tag data}
\usage{
tag(tag, ...)
}
\arguments{
\item{tag}{A tag name}

\item{...}{Named parameters passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Tag data
}
\examples{
\dontrun{
tag('r')
tag('geospatial')
}
}
