% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notifications}
\alias{notifications}
\title{Notifications data}
\usage{
notifications(url = NULL, key = NULL, user = NULL, ...)
}
\arguments{
\item{url}{Base url for a Discourse installation. See 
\code{\link{disc_setup}}}

\item{key}{Your api key on the Discourse installation. See 
\code{\link{disc_setup}}}

\item{user}{Your user name on the Discourse installation. See 
\code{\link{disc_setup}}}

\item{...}{Named parameters passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Notifications data
}
\examples{
\dontrun{
notifications()
}
}
