% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notifications}
\alias{notifications}
\title{Notifications data}
\usage{
notifications(...)
}
\arguments{
\item{...}{Named parameters passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Notifications data
}
\examples{
\dontrun{
notifications()
}
}
