% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/posts.R
\name{posts}
\alias{post_get}
\alias{posts}
\title{Work with posts}
\usage{
post_get(postid, url = NULL, key = NULL, user = NULL, ...)
}
\arguments{
\item{postid}{A post id}

\item{url}{Base url for a Discourse installation. Default is \url{http://discuss.ropensci.org}}

\item{key}{Your api key on the Discourse installation}

\item{user}{Your user name on the Discourse installation}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Work with posts
}
\examples{
\dontrun{
post_get(90)
post_get(120)
post_get(130)
post_get(155)
cat(post_get(155)$raw)
}
#
}

