% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/posts.R
\name{posts}
\alias{get_post}
\alias{posts}
\title{Work with posts}
\usage{
get_post(postid, url = "http://discuss.ropensci.org", key = NULL,
  user = NULL, ...)
}
\arguments{
\item{postid}{A post id}

\item{url}{Base url for a Discourse installation. Default is \url{http://discuss.ropensci.org}}

\item{key}{Your api key on the Discourse installation}

\item{user}{Your user name on the Discourse installation}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Work with posts
}
\examples{
\dontrun{
get_post(90)
get_post(120)
get_post(130)
get_post(155)
cat(get_post(155)$raw)
}
#
}

