% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_messages.R
\name{private_messages}
\alias{private_messages}
\title{Private messages}
\usage{
private_messages(username, ...)
}
\arguments{
\item{username}{A username. required.}

\item{...}{Named parameters passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Private messages
}
\examples{
\dontrun{
private_messages(username = "sckott")
}
}
