% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\alias{reports_page_views}
\title{Reports}
\usage{
reports_page_views(
  start_date,
  end_date,
  category_id = NULL,
  group_id = NULL,
  url = NULL,
  key = NULL,
  ...
)
}
\arguments{
\item{start_date, end_date}{start and end dates}

\item{category_id}{a category id}

\item{group_id}{a group id}

\item{url}{Base url for a Discourse installation. See 
[disc_setup]}

\item{key}{Your api key on the Discourse installation. See 
[disc_setup]}

\item{...}{Named parameters passed on to [crul::verb-GET]}
}
\description{
Reports
}
\examples{
\dontrun{
x <- reports_page_views(start_date = "2019-08-01", end_date = "2019-09-11")
}
}
