% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topics.R
\name{topics}
\alias{latest_topics}
\alias{new_topics}
\alias{topic}
\alias{topics}
\alias{topics_by}
\title{Get the latest topics}
\usage{
latest_topics(url = "http://discuss.ropensci.org", key = NULL,
  user = NULL, ...)

new_topics(url = "http://discuss.ropensci.org", key = NULL, user = NULL,
  ...)

topics_by(id, url = "http://discuss.ropensci.org", key = NULL,
  user = NULL, ...)

topic(id, url = "http://discuss.ropensci.org", key = NULL, user = NULL,
  ...)
}
\arguments{
\item{url}{Base url for a Discourse installation. Default is \url{http://discuss.ropensci.org}}

\item{key}{Your api key on the Discourse installation}

\item{user}{Your user name on the Discourse installation}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}

\item{id}{A user identifier, or a topic identifier. See examples.}
}
\description{
Get the latest topics
}
\examples{
\dontrun{
latest_topics()
new_topics()
topics_by("cboettig")
topic(8)

library("httr")
latest_topics(config=verbose())
}
}

